Added New Script
The "SearchDHCPScopesImproved.ps1" script will allow you to query a Windows DHCP server with the ability to look for an IP Address, MAC Address, or Hostname and find the rest of the corresponding information. This is useful in situations where you may only have a MAC Address of a device and need to hunt down an IP Address or anything else.